\name{ss.data.bolts}
\alias{ss.data.bolts}
\docType{data}
\title{
Example from the book Six Sigma With R. Bolts example.
}
\description{
A data frame with 50 observations of the diameter of the bolts manufactured 
in a production line. 
}

\usage{data(ss.data.bolts)}
\format{
  A data frame with 50 observations on the following variable.
  \describe{
    \item{\code{diameter}}{a numeric vector with the diameter of the bolts}
  }
}
\source{
Data from the authors of the packages.
}

\references{
Cano, Emilio L., Moguerza, Javier M. and Redchuk, Andres. 2012.
\emph{Six Sigma with {R}. Statistical Engineering for Process
  Improvement}, Use R!, vol. 36. Springer, New York.
  \url{http://www.springer.com/statistics/book/978-1-4614-3651-5}.\cr
}
\examples{
data(ss.data.bolts)
}
\keyword{datasets}
\keyword{Loss Function}
