% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ss.data.ca}
\alias{ss.data.ca}
\title{Data for a filling process in a winery}
\format{
A data frame with 20 observations on the following variable.
  \describe{
    \item{\code{Volume}}{a numeric vector (volume in cl}
  }
}
\source{
See references.
}
\usage{
data(ss.data.ca)
}
\description{
The only field of the data is the volume measured in 20 bottles.
}
\note{
This data set is used in chapter 7 of the book ``Six Sigma with R'' (see 
References).
}
\examples{
data(ss.data.ca)
summary(ss.data.ca)
hist(ss.data.ca$Volume)

}
\references{
Cano, Emilio L., Moguerza, Javier M. and Redchuk, Andrés. 2012.
\emph{Six Sigma with {R}. Statistical Engineering for Process
  Improvement}, Use R!, vol. 36. Springer, New York.
  \url{https://link.springer.com/book/10.1007/978-1-4614-3652-2}.\cr
}
\seealso{
\link{ss.study.ca}
}
\keyword{capability}
\keyword{data}
