\name{ss.data.ccConstants}
\alias{ss.data.ccConstants}
\docType{data}
\title{
d2 & d3 constants
}
\description{
Data with constants for control charts limits
}
\usage{data(ss.data.ccConstants)}
\format{
  A data frame with 24 observations on the following 4 variables.
  \describe{
    \item{\code{rr.n}}{Sample Size}
    \item{\code{c4}}{c4 constant}
    \item{\code{d2}}{d2 constant}
    \item{\code{d3}}{d3 constant}
  }
}
\details{
Shewhart's constants for control charts limits computation.
}
\source{
Montgomery, DC (2005). Introduction to Statistical Quality Control
  (Fifth Edition). New York: Wiley&Sons\cr
}
\references{
Montgomery, DC (2005). Introduction to Statistical Quality Control
  (Fifth Edition). New York: Wiley&Sons\cr
}
\examples{
data(ss.data.ccConstants)
ss.data.ccConstants
}
\keyword{datasets}
