\name{ss.data.pb1}
\alias{ss.data.pb1}
\docType{data}
\title{
Particle Boards Example - Individual Data
}
\description{
Humidity of 30 raw material used to make particle boards for individual control chart.
}
\usage{data(ss.data.pb1)}
\format{
  A data frame with 30 observations on the following 2 variables.
  \describe{
    \item{\code{pb.group}}{Group id (distinct for each observation)}
    \item{\code{pb.humidity}}{Humidity of the particle board}
  }
}
\details{
This data are used for the examples in Chapter 12 of the book ``Six Sigma with R''.
}
\source{
E. L. Cano, J. M. Moguerza, and A. Redchuk,
 \emph{Six Sigma with R. Statistical Engineering for Process
 Improvement}, volume 36 of \emph{Use R!}, Springer, New York, 2012.
 ISBN 978-1-4614-3651-5.
 \url{http://www.springer.com/statistics/book/978-1-4614-3651-5}.
}

\references{
Cano, Emilio L., Moguerza, Javier M. and Redchuk, Andres. 2012.
\emph{Six Sigma with {R}. Statistical Engineering for Process
  Improvement}, Use R!, vol. 36. Springer, New York.
  \url{http://www.springer.com/statistics/book/978-1-4614-3651-5}.\cr
}

\examples{
data(ss.data.pb1)
}
\keyword{datasets}
