\name{ss.data.pb2}
\alias{ss.data.pb2}
\docType{data}
\title{
Particle Boards Example - by Groups
}
\description{
Humidity of 20 groups of size 5 of raw materials to make particle boards. For the mean and range control chart.
}
\usage{data(ss.data.pb2)}
\format{
  A data frame with 100 observations on the following 2 variables.
  \describe{
    \item{\code{pb.group}}{a numeric vector}
    \item{\code{pb.humidity}}{a numeric vector}
  }
}
\details{
This data are used for the examples in Chapter 12 of the book ``Six Sigma with R''.
}
\source{
E. L. Cano, J. M. Moguerza, and A. Redchuk,
 \emph{Six Sigma with R. Statistical Engineering for Process
 Improvement}, volume 36 of \emph{Use R!}, Springer, New York, 2012.
 ISBN 978-1-4614-3651-5.
 \url{http://www.springer.com/statistics/book/978-1-4614-3651-5}.
}
\references{
E. L. Cano, J. M. Moguerza, and A. Redchuk,
 \emph{Six Sigma with R. Statistical Engineering for Process
 Improvement}, volume 36 of \emph{Use R!}, Springer, New York, 2012.
 ISBN 978-1-4614-3651-5.
 \url{http://www.springer.com/statistics/book/978-1-4614-3651-5}.
}
\examples{
data(ss.data.pb2)
}
\keyword{datasets}
