\name{ss.data.pb3}
\alias{ss.data.pb3}
\docType{data}
\title{
Particle Boards Example - Attribute data
}
\description{
Counts of raw materials stockouts during 22 weekdays in a month.
}
\usage{data(ss.data.pb3)}
\format{
  A data frame with 22 observations on the following 3 variables.
  \describe{
    \item{\code{day}}{Day id}
    \item{\code{stockouts}}{Number of stockouts}
    \item{\code{orders}}{Number of orders}
  }
}
\details{
This data are used for the examples in Chapter 12 of the book ``Six Sigma with R''.
}
\source{
E. L. Cano, J. M. Moguerza, and A. Redchuk,
 \emph{Six Sigma with R. Statistical Engineering for Process
 Improvement}, volume 36 of \emph{Use R!}, Springer, New York, 2012.
 ISBN 978-1-4614-3651-5.
 \url{http://www.springer.com/statistics/book/978-1-4614-3651-5}.
}

\references{
Cano, Emilio L., Moguerza, Javier M. and Redchuk, Andres. 2012.
\emph{Six Sigma with {R}. Statistical Engineering for Process
  Improvement}, Use R!, vol. 36. Springer, New York.
  \url{http://www.springer.com/statistics/book/978-1-4614-3651-5}.\cr
}

\examples{
data(ss.data.pb3)
}
\keyword{datasets}
