\name{ss.data.pb4}
\alias{ss.data.pb4}
\docType{data}
\title{
Data for Practicle Boards Example - number of defects
}
\description{
Number of defects detected in an order of particle boards.
}
\usage{data(ss.data.pb4)}
\format{
  A data frame with 80 observations on the following 2 variables.
  \describe{
    \item{\code{order}}{Order id}
    \item{\code{defects}}{Number of defects}
  }
}
\details{
This data are used for the examples in Chapter 12 of the book ``Six Sigma with R''.
}
\source{
E. L. Cano, J. M. Moguerza, and A. Redchuk,
 \emph{Six Sigma with R. Statistical Engineering for Process
 Improvement}, volume 36 of \emph{Use R!}, Springer, New York, 2012.
 ISBN 978-1-4614-3651-5.
 \url{http://www.springer.com/statistics/book/978-1-4614-3651-5}.
}
\references{
E. L. Cano, J. M. Moguerza, and A. Redchuk,
 \emph{Six Sigma with R. Statistical Engineering for Process
 Improvement}, volume 36 of \emph{Use R!}, Springer, New York, 2012.
 ISBN 978-1-4614-3651-5.
 \url{http://www.springer.com/statistics/book/978-1-4614-3651-5}.
}
\examples{
data(ss.data.pb4)
}
\keyword{datasets}
