\name{ss.heli}
\alias{ss.heli}
\title{
Creates a pdf file with the design of the Paper Helicopter
}
\description{
The pdf file contains a template with lines and indications to build the 
paper helicopter described in many SixSigma publications.
}
\usage{
ss.heli()
}
\details{
The pdf file must be printed in A4 paper, without adjusting size to paper.
}
\value{
No value is returned
}
\references{
George Box.
Teaching engineers experimental design with a paper helicopter.
\emph{Quality Engineering}, 4(3):453--459, 1992.\cr

Cano, Emilio L., Moguerza, Javier M. and Redchuk, Andres. 2012.
\emph{Six Sigma with {R}. Statistical Engineering for Process
  Improvement}, Use R!, vol. 36. Springer, New York.
  \url{http://www.springer.com/statistics/book/978-1-4614-3651-5}.\cr
}
\author{
Emilio L. Cano
}
\note{
See the vignette("HelicopterInstructions") to see assembling instructions.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
\examples{
ss.heli()
vignette("HelicopterInstructions")
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
