\name{ss.data.bolts}
\alias{ss.data.bolts}
\docType{data}
\title{
Example from the book Six Sigma With R. Bolts example.
}
\description{
A data frame with 50 observations of the diameter of the bolts manufactured 
in a production line. 
}

\usage{data(ss.data.bolts)}
\format{
  A data frame with 50 observations on the following variable.
  \describe{
    \item{\code{diameter}}{a numeric vector with the diameter of the bolts}
  }
}
\source{
Data from the authors of the packages.
}

\examples{
data(ss.data.bolts)
}
\keyword{datasets}
\keyword{Loss Function}
