% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ss.data.pb4}
\alias{ss.data.pb4}
\title{Data for Practicle Boards Example - number of defects}
\format{
A data frame with 80 observations on the following 2 variables.
  \describe{
    \item{\code{order}}{Order id}
    \item{\code{defects}}{Number of defects}
  }
}
\source{
See references.
}
\usage{
data(ss.data.pb4)
}
\description{
Number of defects detected in an order of particle boards.
}
\note{
This data set is used in chapter 12 of the book ``Six Sigma with R'' (see 
References).
}
\examples{
data(ss.data.pb4)
summary(ss.data.pb4)

}
\references{
Cano, Emilio L., Moguerza, Javier M. and Redchuk, Andrés. 2012.
\emph{Six Sigma with {R}. Statistical Engineering for Process
  Improvement}, Use R!, vol. 36. Springer, New York.
  \url{https://www.springer.com/gp/book/9781461436515}.\cr
}
\keyword{cc}
\keyword{data}
