% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ss.data.pc.big}
\alias{ss.data.pc.big}
\title{Larger data set for the printer cartridges example}
\format{A data frame with 72 observations on the following 5 variables, 
  \describe{
    \item{\code{filler}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{batch}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
    \item{\code{col}}{a factor with levels \code{B} \code{C}}
    \item{\code{operator}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{volume}}{a numeric vector}
  }}
\source{
See references.
}
\usage{
data(ss.data.pc.big)
}
\description{
This data set contains data from a simulated process of printer cartridges filling with
complete replications.
}
\note{
This data set is used in chapter 8 of the book ``Six Sigma with R'' (see 
References).
}
\examples{
data(ss.data.pc.big)
summary(ss.data.pc.big)

}
\references{
Cano, Emilio L., Moguerza, Javier M. and Redchuk, Andrés. 2012.
\emph{Six Sigma with {R}. Statistical Engineering for Process
  Improvement}, Use R!, vol. 36. Springer, New York.
  \url{http://www.springer.com/statistics/book/978-1-4614-3651-5}.\cr
}
\keyword{charts}
\keyword{data}

