% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ss.data.pc.r}
\alias{ss.data.pc.r}
\title{Data set for the printer cartridge example, by region}
\format{
A data frame with 5 observations on the following 4 variables.
  \describe{
    \item{\code{pc.regions}}{a factor with levels \code{region.1} \code{region.2} \code{region.3} \code{region.4} \code{region.5}}
    \item{\code{pc.def.a}}{a numeric vector for defects type A}
    \item{\code{pc.def.b}}{a numeric vector for defects type B}
    \item{\code{pc.def}}{a numeric vector for total defects}
  }
}
\source{
See references.
}
\usage{
data(ss.data.pc.r)
}
\description{
This data set contains data from a simulated process of printer cartridge filling. 
The dataframe contains defects by region of each type of cartridge.
}
\note{
This data set is used in chapter 8 of the book ``Six Sigma with R'' (see 
References).
}
\examples{
data(ss.data.pc.r)
summary(ss.data.pc.r)

}
\references{
Cano, Emilio L., Moguerza, Javier M. and Redchuk, Andrés. 2012.
\emph{Six Sigma with {R}. Statistical Engineering for Process
  Improvement}, Use R!, vol. 36. Springer, New York.
  \url{https://link.springer.com/book/10.1007/978-1-4614-3652-2/}.\cr
}
\keyword{charts}
\keyword{data}
