% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ss.data.pc}
\alias{ss.data.pc}
\title{Data set for the printer cartridge example}
\format{
A data frame with 24 observations on the following 6 variables.
  \describe{
    \item{\code{pc.col}}{a factor with levels \code{C} \code{B} for the colour}
    \item{\code{pc.filler}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{pc.volume}}{a numeric vector}
    \item{\code{pc.density}}{a numeric vector}
    \item{\code{pc.batch}}{a numeric vector}
    \item{\code{pc.op}}{a factor with levels \code{A} \code{B} \code{C} \code{D} 
    for the operator}
  }
}
\source{
See references.
}
\usage{
data(ss.data.pc)
}
\description{
This data set contains data from a simulated process of printer cartridge filling.
}
\note{
This data set is used in chapter 8 of the book ``Six Sigma with R'' (see 
References).
}
\examples{
data(ss.data.pc)
summary(ss.data.pc)

}
\references{
Cano, Emilio L., Moguerza, Javier M. and Redchuk, Andrés. 2012.
\emph{Six Sigma with {R}. Statistical Engineering for Process
  Improvement}, Use R!, vol. 36. Springer, New York.
  \url{https://www.springer.com/statistics/book/978-1-4614-3651-5}.\cr
}
\keyword{charts}
\keyword{data}
