% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ss.data.strings}
\alias{ss.data.strings}
\title{Data set for the Guitar Strings example}
\format{
A data frame with 120 observations on the following 6 variables.
  \describe{
    \item{\code{id}}{a numeric vector}
    \item{\code{type}}{a factor with levels \code{A5} \code{B2} \code{D4} \code{E1} \code{E6} \code{G3}}
    \item{\code{res}}{a numeric vector for resistance}
    \item{\code{len}}{a numeric vector for length}
    \item{\code{sound}}{a numeric vector for}
    \item{\code{power}}{a numeric vector}
  }
}
\source{
See references.
}
\usage{
data(ss.data.strings)
}
\description{
This data set contains data from a simulated process of guitar strings production.
}
\note{
This data set is used in chapter 10 of the book ``Six Sigma with R'' (see 
References).
}
\examples{
data(ss.data.strings)
summary(ss.data.strings)

}
\references{
Cano, Emilio L., Moguerza, Javier M. and Redchuk, Andrés. 2012.
\emph{Six Sigma with {R}. Statistical Engineering for Process
  Improvement}, Use R!, vol. 36. Springer, New York.
  \url{https://link.springer.com/book/10.1007/978-1-4614-3652-2/}.\cr
}
\keyword{data}
\keyword{msa}
