% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.lfa.R
\name{ss.lf}
\alias{ss.lf}
\title{Evaluates the Loss Function for a process.}
\usage{
ss.lf(lfa.Y1, lfa.Delta, lfa.Y0, lfa.L0)
}
\arguments{
\item{lfa.Y1}{The observed value of the CTQ (critical to quality) characteristic
that will be evaluated.}

\item{lfa.Delta}{The tolerance for the CTQ.}

\item{lfa.Y0}{The target for the CTQ.}

\item{lfa.L0}{The cost of poor quality when the characteristic is \eqn{Y_0 + \Delta}.}
}
\value{
\item{ss.lf}{A number with the evaluated function at \eqn{Y_1}}
}
\description{
The quality loss function is one of the tools of the Six Sigma methodology.
The function assigns a cost to an observed value, that is larger as far as it
is from the target.
}
\examples{
#Example bolts: evaluate LF at 10.5 if Target=10, Tolerance=0.5, L_0=0.001
ss.lf(10.5, 0.5, 10, 0.001)
}
\references{
Taguchi G, Chowdhury S,Wu Y (2005) \emph{Taguchi's quality engineering handbook}. John
Wiley

Cano, Emilio L., Moguerza, Javier M. and Redchuk, Andres. 2012.
\emph{Six Sigma with {R}. Statistical Engineering for Process
  Improvement}, Use R!, vol. 36. Springer, New York.
  \url{https://link.springer.com/book/10.1007/978-1-4614-3652-2/}.
}
\seealso{
\code{\link{ss.lfa}}
}
\author{
EL Cano
}
\keyword{Taguchi}
\keyword{function}
\keyword{loss}
